Move everything into hats_tap
