+++ src/hats_tap/samples/sample2.adql
