+++ src/hats_tap/samples/sample3.adql
